Added useProductForm.ts
