--- src/components/ProductModal.tsx
@@ -1,1096 +1,0 @@
-import React, { useState, useEffect } from 'react';
-import { X, Plus, Upload, Search, RefreshCw, Package, Settings, GripVertical, Check, ListChecks } from 'lucide-react';
-import { useProductStore } from '../stores/productStore';
-import { useCategoryStore } from '../stores/categoryStore';
-import { useOptionGroupStore, OptionGroup, Option, OptionInput } from '../stores/optionGroupStore';
-import { formatCurrency } from '../utils/formatCurrency';
-import { getActiveCategories, getSortedCategories, getCategoryNameFrequencies, formatCategoryName } from '../utils/categoryUtils';
-import { useToast } from './ui/use-toast';
-import {
-  Dialog,
-  DialogContent,
-  DialogHeader,
-  DialogTitle,
-} from './ui/dialog';
-import { Label } from './ui/label';
-import { Input } from './ui/input';
-import {
-  DndContext,
-  closestCenter,
-  KeyboardSensor,
-  PointerSensor,
-  useSensor,
-  useSensors,
-  type DragEndEvent,
-} from '@dnd-kit/core';
-import {
-  arrayMove,
-  SortableContext,
-  sortableKeyboardCoordinates,
-  useSortable,
-  verticalListSortingStrategy,
-} from '@dnd-kit/sortable';
-import { CSS } from '@dnd-kit/utilities';
-
-interface Props {
-  onClose: () => void;
-}
-
-interface SelectedGroupState {
-  enabled: boolean;
-  options: Record<number, { enabled: boolean; priceOverride: number | null }>;
-}
-
-interface OptionRow {
-  tempId: string;
-  name: string;
-  additional_price: string;
-}
-
-const normalizeText = (text: string): string => {
-  return text.trim().charAt(0).toUpperCase() + text.trim().slice(1).toLowerCase();
-};
-
-const generateTempId = () => Math.random().toString(36).slice(2, 10);
-
-function SortableOptionRow({
-  row,
-  index,
-  onUpdate,
-  onRemove,
-}: {
-  row: OptionRow;
-  index: number;
-  onUpdate: (tempId: string, field: 'name' | 'additional_price', value: string) => void;
-  onRemove: (tempId: string) => void;
-}) {
-  const {
-    attributes,
-    listeners,
-    setNodeRef,
-    transform,
-    transition,
-    isDragging,
-  } = useSortable({ id: row.tempId });
-
-  const style = {
-    transform: CSS.Transform.toString(transform),
-    transition,
-  };
-
-  return (
-    <div
-      ref={setNodeRef}
-      style={style}
-      className={`flex items-center gap-2 group ${isDragging ? 'z-50 opacity-80' : ''}`}
-    >
-      <button
-        type="button"
-        className="flex-shrink-0 p-1 text-gray-300 dark:text-gray-600 hover:text-gray-500 dark:hover:text-gray-400 cursor-grab active:cursor-grabbing touch-none"
-        {...attributes}
-        {...listeners}
-      >
-        <GripVertical className="w-4 h-4" />
-      </button>
-
-      <div className="flex-1 min-w-0">
-        <input
-          type="text"
-          value={row.name}
-          onChange={(e) => onUpdate(row.tempId, 'name', e.target.value)}
-          placeholder={`Opcion ${index + 1}`}
-          maxLength={100}
-          className="w-full px-3 py-2 text-sm rounded-lg border border-gray-200 dark:border-darkbg bg-white dark:bg-darkbg focus:ring-2 focus:ring-primary/20 dark:focus:ring-secondary/20 focus:border-primary dark:focus:border-secondary text-gray-900 dark:text-white placeholder:text-gray-400 dark:placeholder:text-gray-500 transition-colors"
-        />
-      </div>
-
-      <div className="flex-shrink-0 w-24 relative">
-        <span className="absolute left-3 top-1/2 -translate-y-1/2 text-gray-400 dark:text-gray-500 text-sm">$</span>
-        <input
-          type="number"
-          value={row.additional_price}
-          onChange={(e) => onUpdate(row.tempId, 'additional_price', e.target.value)}
-          placeholder="0"
-          min="0"
-          step="0.50"
-          className="w-full pl-7 pr-2 py-2 text-sm rounded-lg border border-gray-200 dark:border-darkbg bg-white dark:bg-darkbg focus:ring-2 focus:ring-primary/20 dark:focus:ring-secondary/20 focus:border-primary dark:focus:border-secondary text-gray-900 dark:text-white placeholder:text-gray-400 dark:placeholder:text-gray-500 transition-colors"
-        />
-      </div>
-
-      <button
-        type="button"
-        onClick={() => onRemove(row.tempId)}
-        className="flex-shrink-0 p-1.5 text-gray-300 dark:text-gray-600 hover:text-red-500 dark:hover:text-red-400 hover:bg-red-50 dark:hover:bg-red-900/20 rounded-lg transition-colors opacity-0 group-hover:opacity-100"
-      >
-        <X className="w-4 h-4" />
-      </button>
-    </div>
-  );
-}
-
-export default function ProductModal({ onClose }: Props) {
-  const { selectedProduct, createProduct, updateProduct } = useProductStore();
-  const { categories, fetchCategories } = useCategoryStore();
-  const {
-    groups: optionGroups,
-    fetchGroups,
-    isGroupModalOpen,
-    setIsGroupModalOpen,
-    saveGroupWithOptions,
-  } = useOptionGroupStore();
-
-  const { toast } = useToast();
-  const [loading, setLoading] = useState(false);
-  const [activeTab, setActiveTab] = useState<'basic' | 'options'>('basic');
-  const [formData, setFormData] = useState({
-    name: selectedProduct?.name || '',
-    category_id: selectedProduct?.category_id || 0,
-    description: selectedProduct?.description || '',
-    base_price: selectedProduct?.base_price || 0,
-    image_url: selectedProduct?.image_url || '',
-    active: selectedProduct?.active ?? true,
-  });
-
-  const [selectedGroups, setSelectedGroups] = useState<Record<number, SelectedGroupState>>({});
-  const [loadingOptions, setLoadingOptions] = useState(false);
-  const [optionSearch, setOptionSearch] = useState('');
-  const [imageFile, setImageFile] = useState<File | null>(null);
-  const [imagePreview, setImagePreview] = useState<string | null>(null);
-  const [error, setError] = useState<string | null>(null);
-
-  const [newGroupData, setNewGroupData] = useState({
-    name: '',
-    min_select: 0,
-    max_select: 1,
-  });
-  const [newGroupOptions, setNewGroupOptions] = useState<OptionRow[]>([
-    { tempId: generateTempId(), name: '', additional_price: '' }
-  ]);
-  const [savingGroup, setSavingGroup] = useState(false);
-  const [groupFormErrors, setGroupFormErrors] = useState<Record<string, string>>({});
-
-  const dndSensors = useSensors(
-    useSensor(PointerSensor, { activationConstraint: { distance: 5 } }),
-    useSensor(KeyboardSensor, { coordinateGetter: sortableKeyboardCoordinates })
-  );
-
-  useEffect(() => {
-    fetchCategories();
-    setLoadingOptions(true);
-    fetchGroups().finally(() => setLoadingOptions(false));
-
-    if (selectedProduct) {
-      if (selectedProduct.image_url) {
-        setImagePreview(selectedProduct.image_url);
-      }
-
-      const initialState: Record<number, SelectedGroupState> = {};
-      selectedProduct.option_groups?.forEach(pog => {
-        initialState[pog.option_group_id] = {
-          enabled: true,
-          options: {},
-        };
-        pog.options.forEach(po => {
-          initialState[pog.option_group_id].options[po.option_id] = {
-            enabled: po.active,
-            priceOverride: po.additional_price,
-          };
-        });
-      });
-      setSelectedGroups(initialState);
-    }
-  }, [fetchCategories, fetchGroups, selectedProduct]);
-
-  useEffect(() => {
-    return () => {
-      setIsGroupModalOpen(false);
-    };
-  }, [setIsGroupModalOpen]);
-
-  const filteredGroups = optionGroups
-    .filter(group =>
-      group.active &&
-      (group.name.toLowerCase().includes(optionSearch.toLowerCase()) ||
-        group.options.some(opt => opt.name.toLowerCase().includes(optionSearch.toLowerCase())))
-    )
-    .sort((a, b) => a.name.localeCompare(b.name));
-
-  const validateForm = () => {
-    if (!formData.name.trim()) {
-      setError('El nombre del producto es requerido');
-      return false;
-    }
-    if (!formData.category_id) {
-      setError('Debe seleccionar una categoría');
-      return false;
-    }
-    if (formData.base_price <= 0) {
-      setError('El precio base debe ser mayor a 0');
-      return false;
-    }
-    if (imageFile && imageFile.size > 2 * 1024 * 1024) {
-      setError('La imagen no debe superar los 2MB');
-      return false;
-    }
-    return true;
-  };
-
-  const handleImageChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const file = e.target.files?.[0];
-    if (file) {
-      if (file.size > 2 * 1024 * 1024) {
-        setError('La imagen no debe superar los 2MB');
-        return;
-      }
-      setImageFile(file);
-      const previewUrl = URL.createObjectURL(file);
-      setImagePreview(previewUrl);
-      setError(null);
-    }
-  };
-
-  const handleGroupToggle = (groupId: number) => {
-    const group = optionGroups.find(g => g.id === groupId);
-    if (!group) return;
-
-    setSelectedGroups(prev => {
-      const current = prev[groupId];
-      if (current?.enabled) {
-        return {
-          ...prev,
-          [groupId]: { ...current, enabled: false },
-        };
-      }
-
-      const activeOptions = group.options.filter(o => o.active);
-      const existingOptions = current?.options ?? {};
-
-      const optionsState: Record<number, { enabled: boolean; priceOverride: number | null }> = {};
-      activeOptions.forEach(opt => {
-        if (existingOptions[opt.id]) {
-          optionsState[opt.id] = existingOptions[opt.id];
-        } else {
-          optionsState[opt.id] = { enabled: true, priceOverride: null };
-        }
-      });
-
-      return {
-        ...prev,
-        [groupId]: { enabled: true, options: optionsState },
-      };
-    });
-  };
-
-  const handleOptionToggle = (groupId: number, optionId: number, option: Option) => {
-    setSelectedGroups(prev => {
-      const groupState = prev[groupId];
-      if (!groupState) return prev;
-
-      const optionState = groupState.options[optionId];
-      const newEnabled = !optionState?.enabled;
-
-      return {
-        ...prev,
-        [groupId]: {
-          ...groupState,
-          options: {
-            ...groupState.options,
-            [optionId]: {
-              enabled: newEnabled,
-              priceOverride: optionState?.priceOverride ?? null,
-            },
-          },
-        },
-      };
-    });
-  };
-
-  const handlePriceOverride = (groupId: number, optionId: number, price: number | null) => {
-    setSelectedGroups(prev => {
-      const groupState = prev[groupId];
-      if (!groupState) return prev;
-
-      return {
-        ...prev,
-        [groupId]: {
-          ...groupState,
-          options: {
-            ...groupState.options,
-            [optionId]: {
-              ...groupState.options[optionId],
-              priceOverride: price,
-            },
-          },
-        },
-      };
-    });
-  };
-
-  const resetNewGroupForm = () => {
-    setNewGroupData({ name: '', min_select: 0, max_select: 1 });
-    setNewGroupOptions([{ tempId: generateTempId(), name: '', additional_price: '' }]);
-    setGroupFormErrors({});
-  };
-
-  const handleCloseGroupModal = () => {
-    setIsGroupModalOpen(false);
-    resetNewGroupForm();
-  };
-
-  const updateGroupOptionRow = (tempId: string, field: 'name' | 'additional_price', value: string) => {
-    setNewGroupOptions(prev => prev.map(r => r.tempId === tempId ? { ...r, [field]: value } : r));
-    if (groupFormErrors.options) setGroupFormErrors(prev => ({ ...prev, options: '' }));
-  };
-
-  const removeGroupOptionRow = (tempId: string) => {
-    setNewGroupOptions(prev => {
-      const filtered = prev.filter(r => r.tempId !== tempId);
-      return filtered.length === 0
-        ? [{ tempId: generateTempId(), name: '', additional_price: '' }]
-        : filtered;
-    });
-  };
-
-  const addGroupOptionRow = () => {
-    setNewGroupOptions(prev => [...prev, { tempId: generateTempId(), name: '', additional_price: '' }]);
-  };
-
-  const handleGroupDragEnd = (event: DragEndEvent) => {
-    const { active, over } = event;
-    if (!over || active.id === over.id) return;
-
-    setNewGroupOptions(prev => {
-      const oldIndex = prev.findIndex(r => r.tempId === active.id);
-      const newIndex = prev.findIndex(r => r.tempId === over.id);
-      return arrayMove(prev, oldIndex, newIndex);
-    });
-  };
-
-  const validateGroupForm = () => {
-    const newErrors: Record<string, string> = {};
-    const normalizedName = normalizeText(newGroupData.name);
-
-    if (!normalizedName) {
-      newErrors.name = 'El nombre es requerido';
-    } else if (normalizedName.length > 100) {
-      newErrors.name = 'El nombre no puede exceder los 100 caracteres';
-    } else {
-      const isDuplicate = optionGroups.some(group =>
-        group.name.toLowerCase() === normalizedName.toLowerCase()
-      );
-      if (isDuplicate) {
-        newErrors.name = 'Ya existe un grupo con este nombre';
-      }
-    }
-
-    if (newGroupData.min_select < 0) {
-      newErrors.min_select = 'No puede ser negativo';
-    }
-    if (newGroupData.max_select < 1) {
-      newErrors.max_select = 'Debe ser al menos 1';
-    }
-    if (newGroupData.min_select > newGroupData.max_select) {
-      newErrors.min_select = 'No puede ser mayor que el maximo';
-    }
-
-    const validOptions = newGroupOptions.filter(r => r.name.trim() !== '');
-    const optionNames = validOptions.map(r => normalizeText(r.name).toLowerCase());
-    const hasDuplicateOptions = new Set(optionNames).size !== optionNames.length;
-    if (hasDuplicateOptions) {
-      newErrors.options = 'Hay opciones con nombres duplicados';
-    }
-
-    for (const row of validOptions) {
-      const price = parseFloat(row.additional_price) || 0;
-      if (price < 0) {
-        newErrors.options = 'Los precios no pueden ser negativos';
-        break;
-      }
-    }
-
-    setGroupFormErrors(newErrors);
-    return Object.keys(newErrors).length === 0;
-  };
-
-  const handleSaveNewGroup = async () => {
-    if (!validateGroupForm()) return;
-
-    try {
-      setSavingGroup(true);
-      const normalizedName = normalizeText(newGroupData.name);
-
-      const validOptions: OptionInput[] = newGroupOptions
-        .filter(r => r.name.trim() !== '')
-        .map(r => ({
-          name: normalizeText(r.name),
-          additional_price: parseFloat(r.additional_price) || 0,
-        }));
-
-      await saveGroupWithOptions({
-        groupId: null,
-        name: normalizedName,
-        min_select: newGroupData.min_select,
-        max_select: newGroupData.max_select,
-        active: true,
-        options: validOptions,
-      });
-
-      setLoadingOptions(true);
-      await fetchGroups();
-      setLoadingOptions(false);
-
-      toast({
-        title: 'Grupo creado',
-        description: 'El grupo y sus opciones se han creado exitosamente',
-      });
-
-      handleCloseGroupModal();
-    } catch (err: unknown) {
-      const message = err instanceof Error ? err.message : 'Error al crear el grupo';
-      toast({
-        variant: 'destructive',
-        title: 'Error',
-        description: message,
-      });
-    } finally {
-      setSavingGroup(false);
-    }
-  };
-
-  const getGroupSelectionDescription = () => {
-    const { min_select, max_select } = newGroupData;
-    if (min_select === 0 && max_select === 1) return 'El cliente puede seleccionar hasta 1 opcion (opcional)';
-    if (min_select === 1 && max_select === 1) return 'El cliente debe seleccionar exactamente 1 opcion (requerido)';
-    if (min_select === 0 && max_select > 1) return `El cliente puede seleccionar hasta ${max_select} opciones (opcional)`;
-    if (min_select > 0 && max_select > 1 && min_select < max_select) return `El cliente debe seleccionar entre ${min_select} y ${max_select} opciones`;
-    if (min_select > 0 && min_select === max_select && max_select > 1) return `El cliente debe seleccionar exactamente ${max_select} opciones`;
-    return '';
-  };
-
-  const validNewGroupOptionCount = newGroupOptions.filter(r => r.name.trim() !== '').length;
-
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-    if (!validateForm()) return;
-
-    try {
-      setLoading(true);
-      setError(null);
-
-      const enabledGroups = Object.entries(selectedGroups)
-        .filter(([_, state]) => state.enabled)
-        .map(([groupId, state]) => {
-          const gId = parseInt(groupId);
-          const group = optionGroups.find(g => g.id === gId);
-          return {
-            groupId: gId,
-            groupName: group?.name || '',
-            state,
-          };
-        })
-        .sort((a, b) => a.groupName.localeCompare(b.groupName));
-
-      const optionGroupsData = enabledGroups.map((item, index) => ({
-        option_group_id: item.groupId,
-        sort_order: index,
-        options: Object.entries(item.state.options)
-          .filter(([_, optState]) => optState.enabled)
-          .map(([optionId, optState]) => ({
-            option_id: parseInt(optionId),
-            additional_price: optState.priceOverride,
-          })),
-      }));
-
-      const productData = {
-        ...formData,
-        option_groups: optionGroupsData,
-      };
-
-      if (selectedProduct) {
-        await updateProduct(selectedProduct.id, productData, imageFile || undefined);
-      } else {
-        await createProduct(productData, imageFile || undefined);
-      }
-
-      toast({
-        title: selectedProduct ? 'Producto actualizado' : 'Producto creado',
-        description: selectedProduct ? 'El producto se ha actualizado exitosamente' : 'El producto se ha creado exitosamente',
-      });
-      onClose();
-    } catch (error: any) {
-      setError(error.message || 'Error al guardar el producto');
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  const tabs = [
-    { id: 'basic', label: 'Información Básica', icon: Package },
-    { id: 'options', label: 'Opciones', icon: Settings },
-  ];
-
-  const activeGroupsCount = Object.values(selectedGroups).filter(g => g.enabled).length;
-
-  return (
-    <Dialog open onOpenChange={onClose}>
-      <DialogContent className="max-w-4xl p-0 overflow-hidden">
-        <div className="flex flex-col h-[85vh]">
-          <DialogHeader className="px-6 py-4 border-b border-gray-200 dark:border-darkbg bg-gradient-to-r from-primary/5 to-secondary/5 dark:from-primary/10 dark:to-secondary/10">
-            <div className="flex items-center justify-between">
-              <DialogTitle className="text-xl font-bold text-primary-dark dark:text-white flex items-center gap-2">
-                <Package className="w-6 h-6" />
-                {selectedProduct ? 'Editar Producto' : 'Nuevo Producto'}
-              </DialogTitle>
-              <button
-                onClick={onClose}
-                className="p-2 hover:bg-gray-100 dark:hover:bg-darkbg rounded-lg transition-colors"
-              >
-                <X className="w-5 h-5 text-gray-500 dark:text-gray-400" />
-              </button>
-            </div>
-          </DialogHeader>
-
-          <div className="px-6 py-3 border-b border-gray-200 dark:border-darkbg bg-gray-50 dark:bg-darkbg-darker">
-            <div className="flex space-x-1">
-              {tabs.map((tab) => {
-                const Icon = tab.icon;
-                const isActive = activeTab === tab.id;
-                let badge = '';
-
-                if (tab.id === 'options' && activeGroupsCount > 0) {
-                  badge = activeGroupsCount.toString();
-                }
-
-                return (
-                  <button
-                    key={tab.id}
-                    onClick={() => setActiveTab(tab.id as any)}
-                    className={`flex items-center gap-2 px-4 py-2 rounded-lg text-sm font-medium transition-all ${
-                      isActive
-                        ? 'bg-white dark:bg-darkbg-lighter text-primary dark:text-secondary shadow-sm'
-                        : 'text-gray-600 dark:text-gray-400 hover:text-gray-900 dark:hover:text-white hover:bg-white/50 dark:hover:bg-darkbg-lighter/50'
-                    }`}
-                  >
-                    <Icon className="w-4 h-4" />
-                    {tab.label}
-                    {badge && (
-                      <span className="ml-1 px-2 py-0.5 text-xs bg-primary dark:bg-secondary text-white rounded-full">
-                        {badge}
-                      </span>
-                    )}
-                  </button>
-                );
-              })}
-            </div>
-          </div>
-
-          <div className="flex-1 overflow-y-auto">
-            <form id="productForm" onSubmit={handleSubmit} className="p-6">
-              {error && (
-                <div className="mb-6 bg-red-50 dark:bg-red-900/20 border border-red-200 dark:border-red-800 rounded-lg p-4 text-red-700 dark:text-red-400">
-                  {error}
-                </div>
-              )}
-
-              {activeTab === 'basic' && (
-                <div className="space-y-6">
-                  <div className="grid grid-cols-1 lg:grid-cols-2 gap-6">
-                    <div className="space-y-4">
-                      <div>
-                        <Label htmlFor="category_id">Categoría</Label>
-                        <select
-                          id="category_id"
-                          value={formData.category_id}
-                          onChange={(e) => setFormData(prev => ({
-                            ...prev,
-                            category_id: parseInt(e.target.value)
-                          }))}
-                          className="w-full px-4 py-3 rounded-lg border border-gray-300 dark:border-darkbg focus:ring-2 focus:ring-primary/20 dark:focus:ring-secondary/20 focus:border-primary dark:focus:border-secondary bg-white dark:bg-darkbg-lighter text-gray-900 dark:text-white"
-                          required
-                        >
-                          <option value="">Seleccionar Categoría</option>
-                          {(() => {
-                            const activeCategories = getActiveCategories(categories);
-                            const sortedCategories = getSortedCategories(activeCategories);
-                            const frequencies = getCategoryNameFrequencies(activeCategories);
-                            return sortedCategories.map(category => (
-                              <option key={category.id} value={category.id}>
-                                {formatCategoryName(category, frequencies)}
-                              </option>
-                            ));
-                          })()}
-                        </select>
-                      </div>
-
-                      <div>
-                        <Label htmlFor="name">Nombre del Producto</Label>
-                        <Input
-                          id="name"
-                          value={formData.name}
-                          onChange={(e) => setFormData(prev => ({ ...prev, name: e.target.value }))}
-                          className="h-12"
-                          placeholder="Ej: Cafe Americano"
-                          required
-                        />
-                      </div>
-
-                      <div>
-                        <Label htmlFor="base_price">Precio Base</Label>
-                        <div className="relative">
-                          <span className="absolute left-3 top-1/2 transform -translate-y-1/2 text-gray-500 dark:text-gray-400">$</span>
-                          <Input
-                            type="number"
-                            id="base_price"
-                            value={formData.base_price}
-                            onChange={(e) => setFormData(prev => ({
-                              ...prev,
-                              base_price: parseFloat(e.target.value) || 0
-                            }))}
-                            min="0"
-                            step="0.50"
-                            className="pl-8 h-12"
-                            placeholder="0.00"
-                            required
-                          />
-                        </div>
-                      </div>
-
-                      <div>
-                        <Label htmlFor="description">Descripción</Label>
-                        <textarea
-                          id="description"
-                          value={formData.description}
-                          onChange={(e) => setFormData(prev => ({ ...prev, description: e.target.value }))}
-                          rows={4}
-                          className="w-full px-4 py-3 rounded-lg border border-gray-300 dark:border-darkbg focus:ring-2 focus:ring-primary/20 dark:focus:ring-secondary/20 focus:border-primary dark:focus:border-secondary bg-white dark:bg-darkbg-lighter text-gray-900 dark:text-white resize-none"
-                          placeholder="Describe tu producto..."
-                        />
-                      </div>
-
-                      <div className="flex items-center gap-3 p-4 bg-gray-50 dark:bg-darkbg-lighter rounded-lg">
-                        <input
-                          type="checkbox"
-                          id="active"
-                          checked={formData.active}
-                          onChange={(e) => setFormData(prev => ({
-                            ...prev,
-                            active: e.target.checked
-                          }))}
-                          className="w-4 h-4 rounded border-gray-300 dark:border-darkbg text-primary dark:text-secondary focus:ring-primary/20 dark:focus:ring-secondary/20"
-                        />
-                        <Label htmlFor="active" className="text-sm font-medium">
-                          Producto Activo
-                        </Label>
-                      </div>
-                    </div>
-
-                    <div>
-                      <Label>Imagen del Producto</Label>
-                      <div className="mt-2 space-y-4">
-                        {imagePreview && (
-                          <div className="relative w-full h-48 rounded-lg overflow-hidden bg-gray-100 dark:bg-darkbg">
-                            <img
-                              src={imagePreview}
-                              alt="Preview"
-                              className="w-full h-full object-cover"
-                            />
-                            <button
-                              type="button"
-                              onClick={() => {
-                                setImagePreview(null);
-                                setImageFile(null);
-                                setFormData(prev => ({ ...prev, image_url: '' }));
-                              }}
-                              className="absolute top-2 right-2 p-1.5 bg-red-600 text-white rounded-full hover:bg-red-700 transition-colors"
-                            >
-                              <X className="w-4 h-4" />
-                            </button>
-                          </div>
-                        )}
-                        <div className="flex items-center justify-center w-full">
-                          <label className="w-full flex flex-col items-center px-6 py-8 bg-white dark:bg-darkbg-lighter rounded-lg border-2 border-gray-300 dark:border-darkbg border-dashed cursor-pointer hover:bg-gray-50 dark:hover:bg-darkbg transition-colors">
-                            <Upload className="w-10 h-10 text-gray-400 dark:text-gray-500 mb-3" />
-                            <span className="text-sm font-medium text-gray-700 dark:text-gray-300">
-                              {imageFile ? imageFile.name : 'Seleccionar imagen'}
-                            </span>
-                            <span className="text-xs text-gray-500 dark:text-gray-400 mt-1">
-                              PNG, JPG hasta 2MB
-                            </span>
-                            <input
-                              type="file"
-                              className="hidden"
-                              accept="image/*"
-                              onChange={handleImageChange}
-                            />
-                          </label>
-                        </div>
-                      </div>
-                    </div>
-                  </div>
-                </div>
-              )}
-
-              {activeTab === 'options' && (
-                <div className="space-y-6">
-                  <div className="flex items-center justify-between">
-                    <div>
-                      <h3 className="text-lg font-semibold text-gray-900 dark:text-white">Grupos de Opciones</h3>
-                      <p className="text-sm text-gray-600 dark:text-gray-400">Selecciona los grupos de opciones disponibles para este producto</p>
-                    </div>
-                    <div className="flex items-center gap-2">
-                      <button
-                        type="button"
-                        onClick={() => {
-                          setLoadingOptions(true);
-                          fetchGroups().finally(() => setLoadingOptions(false));
-                        }}
-                        disabled={loadingOptions}
-                        className="p-2 text-gray-500 dark:text-gray-400 hover:text-primary dark:hover:text-secondary hover:bg-gray-100 dark:hover:bg-darkbg rounded-lg transition-colors"
-                        title="Recargar opciones"
-                      >
-                        <RefreshCw className={`w-5 h-5 ${loadingOptions ? 'animate-spin' : ''}`} />
-                      </button>
-                      <button
-                        type="button"
-                        onClick={() => setIsGroupModalOpen(true)}
-                        className="flex items-center gap-2 px-4 py-2 text-sm font-medium text-white bg-primary dark:bg-secondary rounded-lg hover:bg-primary-dark dark:hover:bg-secondary/90 transition-colors"
-                      >
-                        <Plus className="w-4 h-4" />
-                        Nuevo Grupo
-                      </button>
-                    </div>
-                  </div>
-
-                  <div className="relative">
-                    <Search className="w-5 h-5 absolute left-3 top-1/2 transform -translate-y-1/2 text-gray-400" />
-                    <Input
-                      type="text"
-                      placeholder="Buscar grupos u opciones..."
-                      value={optionSearch}
-                      onChange={(e) => setOptionSearch(e.target.value)}
-                      className="pl-10 h-12"
-                    />
-                  </div>
-
-                  <div className="space-y-4 max-h-[400px] overflow-y-auto">
-                    {loadingOptions ? (
-                      <div className="text-center py-8">
-                        <div className="w-8 h-8 border-2 border-primary dark:border-secondary border-t-transparent rounded-full animate-spin mx-auto" />
-                        <p className="text-sm text-gray-500 dark:text-gray-400 mt-2">Cargando opciones...</p>
-                      </div>
-                    ) : filteredGroups.length === 0 ? (
-                      <div className="text-center py-8">
-                        <Settings className="w-12 h-12 text-gray-400 dark:text-gray-500 mx-auto mb-3" />
-                        <p className="text-sm text-gray-500 dark:text-gray-400">
-                          {optionSearch ? 'No se encontraron grupos' : 'No hay grupos de opciones disponibles'}
-                        </p>
-                      </div>
-                    ) : (
-                      filteredGroups.map(group => {
-                        const groupState = selectedGroups[group.id];
-                        const isGroupEnabled = groupState?.enabled || false;
-                        const activeOptions = group.options.filter(o => o.active);
-
-                        return (
-                          <div
-                            key={group.id}
-                            className={`border rounded-lg transition-all ${
-                              isGroupEnabled
-                                ? 'border-primary/30 dark:border-secondary/30 bg-primary/5 dark:bg-secondary/5'
-                                : 'border-gray-200 dark:border-darkbg bg-white dark:bg-darkbg-lighter'
-                            }`}
-                          >
-                            <div className="p-4">
-                              <div className="flex items-center justify-between">
-                                <div className="flex items-center gap-3">
-                                  <input
-                                    type="checkbox"
-                                    id={`group-${group.id}`}
-                                    checked={isGroupEnabled}
-                                    onChange={() => handleGroupToggle(group.id)}
-                                    className="w-4 h-4 rounded border-gray-300 dark:border-darkbg text-primary dark:text-secondary focus:ring-primary/20 dark:focus:ring-secondary/20"
-                                  />
-                                  <div>
-                                    <label htmlFor={`group-${group.id}`} className="font-medium text-gray-900 dark:text-white cursor-pointer">
-                                      {group.name}
-                                    </label>
-                                    <p className="text-xs text-gray-500 dark:text-gray-400">
-                                      Min: {group.min_select} | Max: {group.max_select} | {activeOptions.length} opciones
-                                    </p>
-                                  </div>
-                                </div>
-                              </div>
-
-                              {isGroupEnabled && activeOptions.length > 0 && (
-                                <div className="mt-4 pt-4 border-t border-gray-200 dark:border-darkbg space-y-3">
-                                  <p className="text-xs font-medium text-gray-500 dark:text-gray-400 uppercase">
-                                    Opciones habilitadas para este producto
-                                  </p>
-                                  <div className="grid grid-cols-1 md:grid-cols-2 gap-3">
-                                    {activeOptions.map(option => {
-                                      const optionState = groupState?.options[option.id];
-                                      const isOptionEnabled = optionState?.enabled || false;
-
-                                      return (
-                                        <div
-                                          key={option.id}
-                                          className={`p-3 border rounded-lg ${
-                                            isOptionEnabled
-                                              ? 'border-primary/20 dark:border-secondary/20 bg-white dark:bg-darkbg'
-                                              : 'border-gray-100 dark:border-darkbg-darker bg-gray-50 dark:bg-darkbg-darker opacity-60'
-                                          }`}
-                                        >
-                                          <div className="flex items-center justify-between mb-2">
-                                            <div className="flex items-center gap-2">
-                                              <input
-                                                type="checkbox"
-                                                id={`option-${group.id}-${option.id}`}
-                                                checked={isOptionEnabled}
-                                                onChange={() => handleOptionToggle(group.id, option.id, option)}
-                                                className="w-3.5 h-3.5 rounded border-gray-300 dark:border-darkbg text-primary dark:text-secondary focus:ring-primary/20 dark:focus:ring-secondary/20"
-                                              />
-                                              <label
-                                                htmlFor={`option-${group.id}-${option.id}`}
-                                                className="text-sm font-medium text-gray-900 dark:text-white cursor-pointer"
-                                              >
-                                                {option.name}
-                                              </label>
-                                            </div>
-                                            <span className="text-xs text-gray-500 dark:text-gray-400">
-                                              Base: {option.additional_price > 0 ? `+${formatCurrency(option.additional_price)}` : 'Sin costo'}
-                                            </span>
-                                          </div>
-
-                                          {isOptionEnabled && (
-                                            <div className="mt-2">
-                                              <label className="text-xs text-gray-500 dark:text-gray-400">
-                                                Precio override (dejar vacio para usar precio base)
-                                              </label>
-                                              <div className="relative mt-1">
-                                                <span className="absolute left-2 top-1/2 transform -translate-y-1/2 text-gray-400 text-xs">$</span>
-                                                <input
-                                                  type="number"
-                                                  value={optionState?.priceOverride ?? ''}
-                                                  onChange={(e) => {
-                                                    const val = e.target.value;
-                                                    handlePriceOverride(
-                                                      group.id,
-                                                      option.id,
-                                                      val === '' ? null : parseFloat(val)
-                                                    );
-                                                  }}
-                                                  min="0"
-                                                  step="0.50"
-                                                  placeholder={option.additional_price.toString()}
-                                                  className="w-full pl-6 pr-2 py-1.5 text-sm border border-gray-200 dark:border-darkbg rounded bg-white dark:bg-darkbg-lighter text-gray-900 dark:text-white"
-                                                />
-                                              </div>
-                                            </div>
-                                          )}
-                                        </div>
-                                      );
-                                    })}
-                                  </div>
-                                </div>
-                              )}
-                            </div>
-                          </div>
-                        );
-                      })
-                    )}
-                  </div>
-                </div>
-              )}
-            </form>
-          </div>
-
-          <div className="px-6 py-4 border-t border-gray-200 dark:border-darkbg bg-gray-50 dark:bg-darkbg-darker">
-            <div className="flex justify-end gap-3">
-              <button
-                type="button"
-                onClick={onClose}
-                className="px-6 py-2 text-sm font-medium text-gray-700 dark:text-gray-300 bg-white dark:bg-darkbg-lighter border border-gray-300 dark:border-darkbg rounded-lg hover:bg-gray-50 dark:hover:bg-darkbg transition-colors"
-                disabled={loading}
-              >
-                Cancelar
-              </button>
-              <button
-                type="submit"
-                form="productForm"
-                disabled={loading}
-                className="px-6 py-2 text-sm font-medium text-white bg-primary dark:bg-secondary rounded-lg hover:bg-primary-dark dark:hover:bg-secondary/90 transition-colors disabled:opacity-50 flex items-center gap-2"
-              >
-                {loading ? (
-                  <>
-                    <div className="w-4 h-4 border-2 border-white border-t-transparent rounded-full animate-spin" />
-                    Guardando...
-                  </>
-                ) : (
-                  <>
-                    <Package className="w-4 h-4" />
-                    {selectedProduct ? 'Guardar Cambios' : 'Crear Producto'}
-                  </>
-                )}
-              </button>
-            </div>
-          </div>
-        </div>
-      </DialogContent>
-
-      {isGroupModalOpen && (
-        <Dialog open onOpenChange={handleCloseGroupModal}>
-          <DialogContent className="max-w-lg p-0 overflow-hidden">
-            <DialogHeader className="px-5 py-4 border-b border-gray-200 dark:border-darkbg">
-              <div className="flex items-center justify-between">
-                <DialogTitle className="text-lg font-semibold text-gray-900 dark:text-white flex items-center gap-2">
-                  <ListChecks className="w-5 h-5 text-primary dark:text-secondary" />
-                  Nuevo Grupo de Opciones
-                </DialogTitle>
-                <button
-                  onClick={handleCloseGroupModal}
-                  className="p-1.5 hover:bg-gray-100 dark:hover:bg-darkbg rounded-lg transition-colors"
-                >
-                  <X className="w-4 h-4 text-gray-500 dark:text-gray-400" />
-                </button>
-              </div>
-            </DialogHeader>
-
-            <div className="p-5 space-y-4 max-h-[60vh] overflow-y-auto">
-              <div>
-                <Label htmlFor="newGroupName">Nombre del Grupo</Label>
-                <Input
-                  id="newGroupName"
-                  value={newGroupData.name}
-                  onChange={(e) => {
-                    setNewGroupData(prev => ({ ...prev, name: e.target.value }));
-                    if (groupFormErrors.name) setGroupFormErrors(prev => ({ ...prev, name: '' }));
-                  }}
-                  placeholder="Ej: Salsas, Tamano, Extras"
-                  maxLength={100}
-                  className={`mt-1.5 ${groupFormErrors.name ? 'border-red-500' : ''}`}
-                />
-                {groupFormErrors.name && <p className="mt-1 text-sm text-red-500">{groupFormErrors.name}</p>}
-              </div>
-
-              <div className="grid grid-cols-2 gap-4">
-                <div>
-                  <Label htmlFor="newGroupMinSelect">Seleccion Minima</Label>
-                  <Input
-                    type="number"
-                    id="newGroupMinSelect"
-                    value={newGroupData.min_select}
-                    onChange={(e) => {
-                      setNewGroupData(prev => ({ ...prev, min_select: parseInt(e.target.value) || 0 }));
-                      if (groupFormErrors.min_select) setGroupFormErrors(prev => ({ ...prev, min_select: '' }));
-                    }}
-                    min="0"
-                    className={`mt-1.5 ${groupFormErrors.min_select ? 'border-red-500' : ''}`}
-                  />
-                  {groupFormErrors.min_select && <p className="mt-1 text-sm text-red-500">{groupFormErrors.min_select}</p>}
-                  <p className="mt-1 text-xs text-gray-500 dark:text-gray-400">0 = opcional</p>
-                </div>
-                <div>
-                  <Label htmlFor="newGroupMaxSelect">Seleccion Maxima</Label>
-                  <Input
-                    type="number"
-                    id="newGroupMaxSelect"
-                    value={newGroupData.max_select}
-                    onChange={(e) => {
-                      setNewGroupData(prev => ({ ...prev, max_select: parseInt(e.target.value) || 1 }));
-                      if (groupFormErrors.max_select) setGroupFormErrors(prev => ({ ...prev, max_select: '' }));
-                    }}
-                    min="1"
-                    className={`mt-1.5 ${groupFormErrors.max_select ? 'border-red-500' : ''}`}
-                  />
-                  {groupFormErrors.max_select && <p className="mt-1 text-sm text-red-500">{groupFormErrors.max_select}</p>}
-                </div>
-              </div>
-
-              <div className="p-3 bg-gray-50 dark:bg-darkbg rounded-lg">
-                <p className="text-sm text-gray-600 dark:text-gray-400">{getGroupSelectionDescription()}</p>
-              </div>
-
-              <div className="border-t border-gray-100 dark:border-darkbg pt-4">
-                <div className="flex items-center justify-between mb-3">
-                  <div className="flex items-center gap-2">
-                    <span className="text-sm font-medium text-gray-900 dark:text-white">Opciones</span>
-                    {validNewGroupOptionCount > 0 && (
-                      <span className="text-xs text-gray-400 dark:text-gray-500 bg-gray-100 dark:bg-darkbg px-2 py-0.5 rounded-full">
-                        {validNewGroupOptionCount}
-                      </span>
-                    )}
-                  </div>
-                </div>
-
-                {groupFormErrors.options && (
-                  <p className="mb-3 text-sm text-red-500">{groupFormErrors.options}</p>
-                )}
-
-                <div className="space-y-2 mb-3">
-                  <DndContext
-                    sensors={dndSensors}
-                    collisionDetection={closestCenter}
-                    onDragEnd={handleGroupDragEnd}
-                  >
-                    <SortableContext
-                      items={newGroupOptions.map(r => r.tempId)}
-                      strategy={verticalListSortingStrategy}
-                    >
-                      {newGroupOptions.map((row, index) => (
-                        <SortableOptionRow
-                          key={row.tempId}
-                          row={row}
-                          index={index}
-                          onUpdate={updateGroupOptionRow}
-                          onRemove={removeGroupOptionRow}
-                        />
-                      ))}
-                    </SortableContext>
-                  </DndContext>
-                </div>
-
-                <button
-                  type="button"
-                  onClick={addGroupOptionRow}
-                  className="flex items-center justify-center gap-2 w-full py-2 border-2 border-dashed border-gray-200 dark:border-darkbg hover:border-primary/40 dark:hover:border-secondary/40 rounded-lg text-sm text-gray-400 dark:text-gray-500 hover:text-primary dark:hover:text-secondary transition-colors"
-                >
-                  <Plus className="w-4 h-4" />
-                  Agregar Opcion
-                </button>
-              </div>
-            </div>
-
-            <div className="px-5 py-4 border-t border-gray-200 dark:border-darkbg bg-gray-50 dark:bg-darkbg-darker flex justify-end gap-3">
-              <button
-                type="button"
-                onClick={handleCloseGroupModal}
-                disabled={savingGroup}
-                className="px-4 py-2 text-sm font-medium text-gray-700 dark:text-gray-300 hover:bg-gray-100 dark:hover:bg-darkbg rounded-lg transition-colors"
-              >
-                Cancelar
-              </button>
-              <button
-                type="button"
-                onClick={handleSaveNewGroup}
-                disabled={savingGroup}
-                className="inline-flex items-center gap-2 px-4 py-2 text-sm font-medium text-white bg-primary dark:bg-secondary rounded-lg hover:bg-primary-dark dark:hover:bg-secondary/90 transition-colors disabled:opacity-50"
-              >
-                {savingGroup ? (
-                  <>
-                    <div className="w-4 h-4 border-2 border-white/30 border-t-white rounded-full animate-spin" />
-                    Guardando...
-                  </>
-                ) : (
-                  <>
-                    <Check className="w-4 h-4" />
-                    Crear Grupo
-                  </>
-                )}
-              </button>
-            </div>
-          </DialogContent>
-        </Dialog>
-      )}
-    </Dialog>
-  );
-}
